requirements.txt
heatshrink2>=0.6.0
pyyaml>=5.3.1
sortedcontainers>=2.1.0
